+++ src/lte/test/lte-test-carrier-aggregation.h
@@ -73,7 +73,7 @@ public:
    * \param sizeTb
    * \param componentCarrierId the component carrier ID
    */
-  void UlScheduling (uint32_t frameNo, uint32_t subframeNo, uint16_t rnti, uint8_t mcs, uint16_t sizeTb, uint8_t componentCarrierId);
+  void UlScheduling (uint32_t frameNo, uint32_t subframeNo, uint16_t rnti, uint8_t mcs, uint32_t sizeTb, uint8_t componentCarrierId);
   /// Write result to file function
   void WriteResultToFile ();
